build: use parallelization except on Windows
This is similar to LizardByte/Sunshine#3361, except now multithreaded
processing is not used on Windows.

Unlike LizardByte/Sunshine#3361, we're now enabling parallelization
for both dot graph generation and other doxygen processing tasks as well
(except on Windows).

To check for Windows, we use the macro `CMAKE_HOST_WIN32` rather than
the more well-known `WIN32`, because what matters here is the host
platform, not the target platform. In most cases, these will be the
same, but the issue with parallelization is related to Windows as a
host, not a target. In other words, using Windows to build the
documentation for a macOS project should still be single-threaded.